@@ -9,32 +9,37 @@
     "prepare": "husky install"
   },
   "dependencies": {
+    "@chakra-ui/react": "^2.3.6",
+    "@chakra-ui/system": "^2.3.0",
+    "@choc-ui/chakra-autocomplete": "^5.1.2",
+    "@emotion/react": "^11.10.5",
+    "@emotion/styled": "^11.10.5",
     "@googlemaps/react-wrapper": "^1.1.34",
     "@googlemaps/typescript-guards": "^2.0.0",
-    "@reach/combobox": "^0.17.0",
-    "fast-equals": "^3.0.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "fast-equals": "3",
+    "framer-motion": "^7.6.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-router-dom": "^6.3.0",
     "use-places-autocomplete": "^4.0.0"
   },
   "devDependencies": {
     "@types/google.maps": "^3.49.0",
     "@types/node": "^18.0.0",
-    "@types/react": "^17.0.45",
-    "@types/react-dom": "^17.0.17",
+    "@types/react": "^18.0.24",
+    "@types/react-dom": "^18.0.8",
     "@typescript-eslint/eslint-plugin": "^5.25.0",
     "@typescript-eslint/parser": "^5.25.0",
-    "@vitejs/plugin-react": "^1.3.0",
+    "@vitejs/plugin-react": "^2.2.0",
     "eslint": "^8.16.0",
     "eslint-plugin-react": "^7.30.0",
-    "eslint-plugin-simple-import-sort": "^7.0.0",
+    "eslint-plugin-simple-import-sort": "^8.0.0",
     "husky": "^8.0.0",
     "lint-staged": "^13.0.2",
     "prettier": "^2.6.2",
     "sass": "^1.52.1",
     "typescript": "^4.6.3",
-    "vite": "^2.9.9",
+    "vite": "^3.2.1",
     "vite-plugin-html": "^3.2.0"
   },
   "lint-staged": {
@@ -44,7 +49,8 @@
     ]
   },
   "engines": {
-    "node": "16.x",
-    "npm": "8.x"
-  }
+    "node": "18.x",
+    "yarn": "3.x"
+  },
+  "packageManager": "yarn@3.2.4"
 }
